Add support of Mysql charset param
NodeJS Mysql has support for "charset" property, but Caminte has no, so I just add this to my local version and it works well. This works fine but isn't a full solution when you need to switch from utf8 to utf8mb4 (full utf8, for ex. when need to save emoji in your Mysql table).